+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Change log for ITGSerialTerminalTools
 
 ## Unreleased
+
+## 0.2.0
+
+- new features:
+
+  - function Wait-ExpectedMessage
+  - function Invoke-RemoteCommand